perf(array): remove all mixins in favor of ES6 classes, ~20% faster in basic benchmarks
Fix #7798
