+++ lib/types/documentarray.js
@@ -6,7 +6,6 @@
 
 const CoreMongooseArray = require('./core_array');
 const Document = require('../document');
-const MongooseArray = require('./array');
 const ObjectId = require('./objectid');
 const castObjectId = require('../cast/objectid');
 const getDiscriminatorByValue = require('../queryhelpers').getDiscriminatorByValue;
@@ -59,16 +58,8 @@ function MongooseDocumentArray(values, path, doc) {
   }
   arr[arrayPathSymbol] = path;
 
-  // Values always have to be passed to the constructor to initialize, since
-  // otherwise MongooseArray#push will mark the array as modified to the parent.
-  const keysMA = Object.keys(MongooseArray.mixin);
-  let numKeys = keysMA.length;
-  for (let j = 0; j < numKeys; ++j) {
-    arr[keysMA[j]] = MongooseArray.mixin[keysMA[j]];
-  }
-
   const keysMDA = Object.keys(MongooseDocumentArray.mixin);
-  numKeys = keysMDA.length;
+  let numKeys = keysMDA.length;
   for (let i = 0; i < numKeys; ++i) {
     arr[keysMDA[i]] = MongooseDocumentArray.mixin[keysMDA[i]];
   }
